Change buildsystem to automake
Change buildsystem from custom GNUmakefile to standard automake.

This allows builds with the standard invocation of

    autoreconf -if      # if not building from dist tarball
    ./configure
    make
    sudo make install

The configure.ac is based on the output of autoscan, with quite a
bit of manual work added on top.

Uses the ancient serial-tests test harness in order to re-use the
existing run-tests script.