CREDITS.md: Add heading to help pandoc find a title
Note that we are using the "# Title" form of writing a Markdown
heading even though it is less nice to read as text than the
form with a ====== underlining the title.

The reason is to work around git confusing the ======= together
with the Markdown block quotes below ("> ") with leftover conflict
markers and refusing to commit.